+++ keyboards/nullbitsco/nibble/README.md
@@ -10,6 +10,6 @@ A unique, tweakable 65% keyboard kit built by nullbits. [More info at nullbits.c
 
 Make example for this keyboard (after setting up your build environment):
 
-    make nullbitsco/nibble
+    make nullbitsco/nibble:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
