+++ keyboards/nullbitsco/nibble/bitc_led.c
@@ -1,0 +1,22 @@
+#include "bitc_led.h"
+
+void set_bitc_LED(uint8_t mode) {
+    switch(mode) {
+        case LED_ON:
+            setPinOutput(PIN_LED);
+            writePin(PIN_LED, GPIO_STATE_HIGH);
+        break;
+
+        case LED_DIM:
+            setPinInput(PIN_LED);
+        break;
+
+        case LED_OFF:
+            setPinOutput(PIN_LED);
+            writePin(PIN_LED, GPIO_STATE_LOW);
+        break;
+
+        default:
+        break;
+    }
+}
