--- keyboards/nullbitsco/nibble/keymaps/via/nibble.json
@@ -1,29 +1,0 @@
-{
-  "name": "nibble",
-  "vendorId": "0x6E61",
-  "productId": "0x6060",
-  "lighting": "qmk_rgblight",
-  "customFeatures": ["rotary-encoder"],
-    "customKeycodes": [
-    {"name": "Program Keyboard", "title": "Program Keyboard", "shortName": "Prog"},
-    {"name": "Discord Mute", "title": "Discord Mute", "shortName": "Disc Mute"},
-    {"name": "Discord Deafen", "title": "Discord Deafen", "shortName": "Disc Deaf"},
-    {"name": "Super Alt Tab", "title": "Super Alt Tab", "shortName": "S.A.T."}
-  ],
-  "matrix": {"rows": 5,"cols": 16},
-  "layouts": {
-    "labels": [
-      ["Left Modifiers", "3x1.25U", "2x1.5U"],
-      ["Spacebar", "6.25U/7U", "Split 2.75U | 1.25U | 2.25U"],
-      ["Right Modifiers", "3x1U", "2x1.5U"]
-    ],
-    "keymap": [
-      [{"x":1, "c": "#b0ebff", "a": 7},"0,1",{"c": "#cccccc"},"0,2","0,3","0,4","0,5","0,6","0,7","0,8","0,9","0,10","0,11","0,12","0,13",{"w":2, "c": "#b0ebff"},"0,14","0,15"],
-      [{"c":"#7d9da8"},"1,0",{"w":1.5,"c": "#b0ebff"},"1,1",{"c": "#cccccc"},"1,2","1,3","1,4","1,5","1,6","1,7","1,8","1,9","1,10","1,11","1,12","1,13",{"w":1.5,"c": "#b0ebff"},"1,14","1,15"],
-      [{"c":"#7d9da8"},"2,0",{"w":1.75,"c": "#b0ebff"},"2,1",{"c": "#cccccc"},"2,2","2,3","2,4","2,5","2,6","2,7","2,8","2,9","2,10","2,11","2,12",{"w":2.25, "c": "#b0ebff"},"2,14","2,15"],
-      [{"c":"#7d9da8"},"3,0",{"w":2.25,"c": "#b0ebff"},"3,1",{"c": "#cccccc"},"3,2","3,3","3,4","3,5","3,6","3,7","3,8","3,9","3,10","3,11",{"w":1.75, "c": "#b0ebff"},"3,12","3,14","3,15"],
-      [{"c":"#7d9da8"},"4,0",{"w":1.25,"c": "#b0ebff"},"4,1\n\n\n0,0",{"w":1.25},"4,2\n\n\n0,0",{"w":1.25},"4,3\n\n\n0,0",{"w":6.25, "c": "#cccccc"},"4,8\n\n\n1,0",{"c": "#b0ebff"},"4,9\n\n\n2,0","4,10\n\n\n2,0","4,11\n\n\n2,0","4,12","4,14","4,15"],
-      [{"x":1,"w":1.5},"4,1\n\n\n0,1",{"w":1.5},"4,2\n\n\n0,1",{"x":0.75,"w":2.75, "c": "#cccccc"},"4,8\n\n\n1,1",{"w":1.25},"4,9\n\n\n1,1",{"w":2.25},"4,8\n\n\n1,1",{"w":1.5, "c": "#b0ebff"},"4,10\n\n\n2,1",{"w":1.5},"4,11\n\n\n2,1"]
-    ]
-  }
-}
