+++ keyboards/nullbitsco/nibble/rules.mk
@@ -1,6 +1,3 @@
-# Project specific files
-SRC += matrix.c
-
 # MCU name
 MCU = atmega32u4
 
@@ -40,34 +37,39 @@ F_USB = $(F_CPU)
 # Interrupt driven control endpoint task(+60)
 OPT_DEFS += -DINTERRUPT_CONTROL_ENDPOINT
 
-
-# Boot Section Size in *bytes*
-#   Teensy halfKay   512
-#   Teensy++ halfKay 1024
-#   Atmel DFU loader 4096
-#   LUFA bootloader  4096
-#   USBaspLoader     2048
-OPT_DEFS += -DBOOTLOADER_SIZE=4096
-
+# Bootloader selection
+#   Teensy       halfkay
+#   Pro Micro    caterina
+#   Atmel DFU    atmel-dfu
+#   LUFA DFU     lufa-dfu
+#   QMK DFU      qmk-dfu
+#   ATmega32A    bootloadHID
+#   ATmega328P   USBasp
+BOOTLOADER = atmel-dfu
 
 # Build Options
 #   change yes to no to disable
 #
-BOOTMAGIC_ENABLE = no      	# Virtual DIP switch configuration(+1000)
-MOUSEKEY_ENABLE = no       	# Mouse keys(+4700)
-EXTRAKEY_ENABLE = yes       # Audio control and System control(+450)
-# CONSOLE_ENABLE = yes        # Console for debug(+400)
+BOOTMAGIC_ENABLE = no      	# Virtual DIP switch configuration
+MOUSEKEY_ENABLE = no       	# Mouse keys
+EXTRAKEY_ENABLE = yes       # Audio control and System control
+CONSOLE_ENABLE = no         # Console for debug
 COMMAND_ENABLE = no         # Commands for debug and configuration
-CUSTOM_MATRIX = yes         # Custom matrix file
 SLEEP_LED_ENABLE = no       # Breathing sleep LED during USB suspend
 NKRO_ENABLE = yes           # USB Nkey Rollover
 BACKLIGHT_ENABLE = no     	# Enable keyboard backlight functionality
-MIDI_ENABLE = no            # MIDI support (+2400 to 4200, depending on config)
-UNICODE_ENABLE = no         # Unicode
+RGBLIGHT_ENABLE = yes       # Enable keyboard RGB underglow
+MIDI_ENABLE = no            # MIDI support
 BLUETOOTH_ENABLE = no       # Enable Bluetooth with the Adafruit EZ-Key HID
 AUDIO_ENABLE = no           # Audio output on port C6
 FAUXCLICKY_ENABLE = no      # Use buzzer to emulate clicky switches
-ENCODER_ENABLE = yes		# Enable encoder
-RGBLIGHT_ENABLE = yes
-LTO_ENABLE = yes
-# OLED_DRIVER_ENABLE = yes
+ENCODER_ENABLE = yes		# Use rotary encoder
+LTO_ENABLE = yes			# Link-time optimization
+CUSTOM_MATRIX = lite 		# Lite custom matrix 
+
+# Project specific files
+SRC += matrix.c
+SRC += bitc_led.c
+SRC += big_led.c
+SRC += remote_kb.c
+SRC += tmk_core/common/uart.c
